requirements.txt
Flask==3.1.2
Werkzeug==3.1.5
pyjwt==2.10.1


flask-swagger-ui==4.11.1
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.32.5
pydantic==2.12.5
pydantic-settings==2.12.0
bleach==6.3.0
Flask-WTF==1.2.2
bcrypt==5.0.0